fixed select format & commented out update/delete for testing

+++ albums_migration.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS albums (
     artist VARCHAR(100) NOT NULL,
     name VARCHAR(100) NOT NULL,
     release_date INT NOT NULL,
-    sales FLOAT NOT NULL,
+    sales DECIMAL NOT NULL,
     genre VARCHAR(100),
     PRIMARY KEY (id)
 )
